requirement.txt
streamlit
numpy
pandas
matplotlib
seaborn
plotly
scikit-learn
umap-learn
